add map_data method to skatepark and user models, corresponding specs
